+++ app.py
@@ -16,25 +16,6 @@ def resetar_dados():
     turmas = {}
     professores = {}
     return jsonify({"mensagem": "Dados resetados com sucesso!"}), 200
-
-
-#### ROTA PROFESSORES ####
-class Professores:
-    def __init__(self, id, nome, idade, materia):
-        self.id = id
-        self.nome = nome
-        self.idade = idade
-        self.materia = materia
-
-        professores[id] = {
-            "id": self.id,
-            "nome": self.nome,
-            "idade": self.idade,
-            "materia": self.materia
-        }
-    
-    def serialize(self):
-        return professores[self.id]
     
 # POST PROFESSORES
 @app.route('/professores', methods=['POST'])
