+++ models/professores.py
@@ -1,0 +1,18 @@
+from app import professores
+
+class Professores:
+    def __init__(self, id, nome, idade, materia):
+        self.id = id
+        self.nome = nome
+        self.idade = idade
+        self.materia = materia
+
+        professores[id] = {
+            "id": self.id,
+            "nome": self.nome,
+            "idade": self.idade,
+            "materia": self.materia
+        }
+    
+    def serialize(self):
+        return professores[self.id]
